+++ R/hybrid.R
@@ -44,9 +44,10 @@
 #' and \code{n2i}). Pearson correlation coefficients can be analyzed by supplying 
 #' \code{ri} and \code{ni} to the \code{hybrid} function. These correlation 
 #' coefficients are internally transformed to Fisher's z correlations before 
-#' analyzing the data. The results in the output are of the Fisher's z transformed
-#' correlations. It is also possible to specify the standardized effect sizes and 
-#' its sampling variances directly via the \code{yi} and \code{vi} arguments. 
+#' analyzing the data. The results in the output are those of the Fisher's z 
+#' transformed correlations. It is also possible to specify the standardized 
+#' effect sizes and its sampling variances directly via the \code{yi} and \code{vi} 
+#' arguments. 
 #'  
 #' Two other arguments that need to be specified are \code{side} and \code{conventional}. 
 #' \code{side} indicates whether the effect size in the conventional study was expected
@@ -156,7 +157,7 @@
 #' \code{control = list(verbose = TRUE)} Default values are used if an element is 
 #' not specified. The following elements can be specified by the user:
 #' 
-#' \itemize{
+#' \describe{
 #' \item{\code{int:}}{ A vector of length two that indicates the lower and upper 
 #' bound of the interval that is used for estimating the effect size. The effect 
 #' size estimate should be included in this interval. Its default values are -10
@@ -261,7 +262,7 @@
 #' tobs <- qt(pval/2, df = n1i+n2i-2, lower.tail = FALSE) # Observed t-values 
 #' 
 #' ### Apply hybrid method using the implementation of van Aert and van Assen (2018)
-#' hybrid(tobs = tobs, n1i = n1i, n2i = n2i, side = "right", conventional = c(1,0), 
+#' hybrid(tobs = tobs, n1i = n1i, n2i = n2i, side = "right",
 #' control = list(implementation = "two"))
 #'  
 #' @export
@@ -274,7 +275,7 @@ hybrid <- function(m1i, m2i, mi, ri, sd1i, sd2i, sdi, n1i, n2i, ni, tobs, yi, vi
   # no visible binding for global variable est.fe etc.
   est.fe <- ci.ub.fe <- ci.lb.fe <- zval.fe <- NULL 
   
-  if (!missing("mi") | !missing("tobs") | !missing("m1i") | !missing("ri"))
+  if (missing("conventional"))
   { # If only two studies are supplied as input
     
     if (!missing("mi") & !missing("ni") & !missing("sdi")) 
@@ -329,6 +330,13 @@ hybrid <- function(m1i, m2i, mi, ri, sd1i, sd2i, sdi, n1i, n2i, ni, tobs, yi, vi
       es$conventional <- c(1, 0) # Create variable to indicate that first study is an conventional study
       
       res.repl <- repl(es = es, measure = measure, side = side)
+    } else if (!missing("yi") & !missing("vi"))
+    {
+      measure <- "SPE"
+      es <- escompute(yi = yi, vi = vi, alpha = alpha/2, side = side,
+                      measure = measure)
+      
+      es$conventional <- c(1, 0) # Create variable to indicate that first study is an conventional study
     }
     
     if (es$pval[1] > alpha/2) 
@@ -448,6 +456,18 @@ hybrid <- function(m1i, m2i, mi, ri, sd1i, sd2i, sdi, n1i, n2i, ni, tobs, yi, vi
   
   ##############################################################################
   
+  ### The "conventional" argument needs to be specified with the "multiple" 
+  # implementation. Only with the "two" implementation can the "conventional"
+  # argument be omitted.
+  if (missing("conventional") &
+      con$implementation == "multiple" & 
+      k > 2)
+  {
+    stop("Argument 'conventional' needs to be specified when using the default implementation")
+  }
+  
+  ##############################################################################
+  
   ### In the absence of moderators, fit an intercept-only model
   if (is.null(mods)) 
   { 
@@ -495,35 +515,13 @@ hybrid <- function(m1i, m2i, mi, ri, sd1i, sd2i, sdi, n1i, n2i, ni, tobs, yi, vi
     res4 <- fe_ma(yi = es$yi, vi = es$vi)
     
     ### Transform results of fixed-effect meta-analysis
-    if (measure == "COR") 
-    { # Back transform Fisher z to correlation
-      res4$est.fe <- (exp(2*res4$est.fe) - 1)/(exp(2*res4$est.fe) + 1)
-      res4$ci.lb.fe <- (exp(2*res4$ci.lb.fe) - 1)/(exp(2*res4$ci.lb.fe) + 1)
-      res4$ci.ub.fe <- (exp(2*res4$ci.ub.fe) - 1)/(exp(2*res4$ci.ub.fe) + 1)
-      res4$zval.fe <- res4$zval.fe
-      
-      if (side == "left") 
-      { # Re-mirror estimates
-        res4$est.fe <- est.fe*-1
-        tmp <- ci.ub.fe
-        res4$ci.ub.fe <- ci.lb.fe*-1
-        res4$ci.lb.fe <- tmp*-1
-        res4$zval.fe <- zval.fe*-1
-      }
-      
-    } else if (side == "left" & measure != "COR") 
+    if (side == "left") 
     { # Re-mirror estimates
-      res4$est.fe <- res4$est.fe*-1
-      tmp <- res4$ci.ub.fe
-      res4$ci.ub.fe <- res4$ci.lb.fe*-1
+      res4$est.fe <- est.fe*-1
+      tmp <- ci.ub.fe
+      res4$ci.ub.fe <- ci.lb.fe*-1
       res4$ci.lb.fe <- tmp*-1
-      res4$zval.fe <- res4$zval.fe*-1
-    } else 
-    {
-      res4$est.fe <- res4$est.fe
-      res4$ci.ub.fe <- res4$ci.ub.fe
-      res4$ci.lb.fe <- res4$ci.lb.fe
-      res4$zval.fe <- res4$zval.fe
+      res4$zval.fe <- zval.fe*-1
     }
   } else
   { # Assign NAs to objects if implementation == "multiple"
@@ -538,7 +536,7 @@ hybrid <- function(m1i, m2i, mi, ri, sd1i, sd2i, sdi, n1i, n2i, ni, tobs, yi, vi
                        Qstat = NA, Qpval = NA)
     
     res.repl <- data.frame(est.repl = NA, se.repl = NA, ci.lb.repl = NA, 
-               ci.ub.repl = NA, stat.repl = NA, pval.repl = NA, pval.o = NA)
+                           ci.ub.repl = NA, stat.repl = NA, pval.repl = NA, pval.o = NA)
   }
   
   x <- list(con = con, var_names = var_names, est = res1$est, tau2 = res1$tau2,